reduce number of connections when building auction house routes
