+++ actions/wowToken.js
@@ -1,0 +1,7 @@
+'use server';
+
+import getAccessToken from '../utils/db/getAccessToken';
+
+export async function getWowToken() {
+	const accessToken = await getAccessToken();
+}
