+++ pages/index.jsx
@@ -75,7 +75,7 @@ export default function Home({ data }) {
 			</Head>
 			<Navbar />
 			<main className="flex flex-wrap justify-center items-center gap-[2.5%] bg-white/70 backdrop-blur-md font-bold h-screen lg:w-6/12 w-9/12 overflow-y-scroll text-white pt-5 pl-2.5 pr-2.5 pb-12 text-normal-1 scrollbar-none">
-				{/* <Banner /> */}
+				<Banner />
 				{realmsArr}
 			</main>
 			<Footer />
